% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_edgelist}
\alias{as_edgelist}
\title{Converts a nctx graph to an edgelist}
\usage{
as_edgelist(g)
}
\arguments{
\item{g}{A nctx graph}
}
\value{
An edgelist in the form of an \eqn{|V|\times{2}} matrix
}
\description{
Transform a nctx graph to an edgelist, i.e. a matrix of \eqn{|V|\times{2}} with each row containing an edge by its start and target vertex.
}
