% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{copy_from_igraph}
\alias{copy_from_igraph}
\title{Create a graph as a copy of an igraph graph.}
\usage{
copy_from_igraph(g_i)
}
\arguments{
\item{g_i}{An igraph graph}
}
\value{
The copied nctx graph
}
\description{
This creates a nctx-graph as a copy of an igraph-graph.
}
\details{
Internally, this is a call to
  \code{ends(g_i, E(g_i), names = F)}
with \code{g_i} being an igraph graph.

Note that only structural information will be copied, i.e. attributes like vertex names or edge weights are not copied.
}
\examples{
library(RUnit)
library(igraph)
library(nctx)
file <- system.file("extdata", "bcde.directed.graphml", package="nctx")
g_i <- read_graph(file, "graphml")
checkTrue(igraph::is_directed(g_i))
g <- copy_from_igraph(g_i)
checkTrue(nctx::is_directed(g))
checkEquals(vcount(g_i),  10)
checkEquals(num_vertices(g),  10)
checkEquals(ecount(g_i),  14)
checkEquals(num_edges(g),  14)
checkEquals(igraph::as_edgelist(g_i, names = FALSE), nctx::as_edgelist(g))

file <- system.file("extdata", "bcde.undirected.graphml", package="nctx")
g_i <- read_graph(file, "graphml")
checkTrue(!igraph::is_directed(g_i))
g <- copy_from_igraph(g_i)
checkTrue(!nctx::is_directed(g))
checkEquals(vcount(g_i),  10)
checkEquals(num_vertices(g),  10)
checkEquals(ecount(g_i),  14)
checkEquals(num_edges(g),  14)
checkEquals(igraph::as_edgelist(g_i, names = FALSE), nctx::as_edgelist(g))
}
