% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{create_graph}
\alias{create_graph}
\title{Creates a graph}
\usage{
create_graph(directed = FALSE)
}
\arguments{
\item{directed}{Choose if the graph is directed or not}
}
\value{
A nctx graph
}
\description{
This creates an empty graph.
}
\examples{
library(RUnit)
library(nctx)
g <- create_graph(directed=TRUE)
checkTrue(is_directed(g))
checkEquals(num_vertices(g), 0)
checkEquals(num_edges(g), 0)
g <- create_graph(TRUE)
checkTrue(is_directed(g))
checkEquals(num_vertices(g), 0)
checkEquals(num_edges(g), 0)
g <- create_graph(directed=FALSE)
checkTrue(!is_directed(g))
checkEquals(num_vertices(g), 0)
checkEquals(num_edges(g), 0)
g <- create_graph(FALSE)
checkTrue(!is_directed(g))
checkEquals(num_vertices(g), 0)
checkEquals(num_edges(g), 0)
}
