% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modgraph.R
\name{get_vertex_attribute}
\alias{get_vertex_attribute}
\title{Get a vertex attribute}
\usage{
get_vertex_attribute(g, name, vertex)
}
\arguments{
\item{g}{A nctx graph}

\item{name}{Name of the attribute}

\item{vertex}{ID of the vertex of interest}
}
\value{
The attribute value for the vertex
}
\description{
This retrieves a single attribute for an individual vertex.
}
\examples{
library(RUnit)
library(nctx)
file <- system.file("extdata", "bcde.ctx.dir.graphml", package="nctx")
g <- load_graph(file, directed = TRUE)
checkEquals(get_vertex_attributes(g, "distance_to_A"), c(0.0000000,0.2938160,1.0000000,0.4637440,0.3066590,0.1263650,0.3882870,0.0927837,0.2409850,0.5995490))

checkEquals(get_vertex_attribute(g, "distance_to_A", 3), 1.0)
}
\seealso{
\code{\link{set_vertex_attributes}}

\code{\link{get_vertex_attribute_names}}
}
