% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modgraph.R
\name{get_vertex_attribute_names}
\alias{get_vertex_attribute_names}
\title{Get the available vertex attributes}
\usage{
get_vertex_attribute_names(g)
}
\arguments{
\item{g}{A nctx graph}
}
\value{
The names of the graph attributes
}
\description{
This retrieves a list of vertex attribute names.
}
\details{
Use these names in \code{\link{get_vertex_attribute}} or \code{\link{get_vertex_attributes}} later.
}
\examples{
library(RUnit)
library(nctx)
file <- system.file("extdata", "bcde.ctx.dir.graphml", package="nctx")
g <- load_graph(file, directed = TRUE)
prop_names <- get_vertex_attribute_names(g)
checkTrue(all(prop_names \%in\% c("context","distance_to_A","distance_to_B","distance_to_C","distance_to_D","distance_to_E","distance_to_F","distance_to_G","distance_to_H","distance_to_I","distance_to_J","name")))
}
\seealso{
\code{\link{set_vertex_attributes}}

\code{\link{get_vertex_attributes}}

\code{\link{get_vertex_attribute}}
}
