% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_directed}
\alias{is_directed}
\title{Check if a graph is directed}
\usage{
is_directed(g)
}
\arguments{
\item{g}{A nctx graph}
}
\value{
\code{TRUE} if \code{g} is directed
}
\description{
Check if a graph is directed
}
