% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nctxgraph}
\alias{is_nctxgraph}
\title{Checks for a given object if it is a nctx graph}
\usage{
is_nctxgraph(g)
}
\arguments{
\item{g}{An object to check}
}
\value{
\code{TRUE} if \code{g} contains a nctx graph
}
\description{
Checks for a given object if it is a nctx graph
}
