% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_edges}
\alias{num_edges}
\title{Get the number of edges of a graph}
\usage{
num_edges(g)
}
\arguments{
\item{g}{A nctx graph}
}
\value{
The number of edges of \code{g}
}
\description{
Get the number of edges of a graph
}
