% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_vertices}
\alias{num_vertices}
\title{Get the number of vertices of a graph}
\usage{
num_vertices(g)
}
\arguments{
\item{g}{A nctx graph}
}
\value{
The number of vertices of \code{g}
}
\description{
Get the number of vertices of a graph
}
