\name{nctx-package}
\alias{nctx-package}
\alias{nctx}
\docType{package}
\title{
  nctx: Networks in ConTeXt
}
\description{
  The 'nctx' package provides functionality to obtain shortest paths and infer shortest-path-based centrality metrics in attributed networks under contextual constraints. Enforcing constraints is done by user-defined functions at each step of path iteration.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Mirco Schoenfeld <mirco.schoenfeld@uni-bayreuth.de>
}
\references{
  Citation to tech report goes here!
}
\keyword{ network analysis, attributed network, context, shortest-path, centrality }
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
